requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-dotenv==1.0.0
requests==2.31.0
python-decouple==3.8
openai>=1.0.0
pydantic>=2.0.0
pytz==2023.3
